refactor(gui): stage 4 — extract daemon events + banners modules
- src/app/events.js: bell/attention helpers (onSessionBell,
  clearAttention exported — SessionTerm and setActive still call them
  from main.js), processAliveTransition, and every EventsOn handler
  (project:*, session:*, pty:*, control:disconnect, bell-click,
  control:error) registered inside wireDaemonEvents(deps). Injected:
  switchTo, renderMinimizedTray, updateAppTitle, focus calls, the
  scroll tracer, and banners' isDaemonRestarting.
- src/app/banners.js: daemon-stale + update banners, manualUpdateCheck
  (menu entry point), boot-time update poll.

main.js: 2526 → 2004 lines. Suites: vitest 149, mock e2e 52(+1 skip),
e2e-real 5.
